revise formatoption layout
With this commit, we introduce a more general way
of positioning the widgets in the plotmethod tabs
that control the formatoptions. The layout of
plotmethod tabs is now based on a qgridlayout with
a standard interface that manages the various rows.

Furthermore, we add a combo box to select the plot type
for mapplot and plot2d.